Refactoring TimerBase
* Decoupling `TimerBase`, `Timer`, `Every` and `Animation` from Ruby and
Canvas.
* Shoes3/shoes3#333
