+++ shoes/native.h
@@ -87,8 +87,6 @@ SHOES_CONTROL_REF shoes_native_check(VALUE, shoes_canvas *, shoes_place *, VALUE
 VALUE shoes_native_check_get(SHOES_CONTROL_REF);
 void shoes_native_check_set(SHOES_CONTROL_REF, int);
 SHOES_CONTROL_REF shoes_native_radio(VALUE, shoes_canvas *, shoes_place *, VALUE, VALUE);
-void shoes_native_timer_remove(shoes_canvas *, SHOES_TIMER_REF);
-SHOES_TIMER_REF shoes_native_timer_start(VALUE, shoes_canvas *, unsigned int);
 void shoes_native_canvas_oneshot(int, VALUE);
 VALUE shoes_native_clipboard_get(shoes_app *);
 void shoes_native_clipboard_set(shoes_app *, VALUE);
